Crate faster_hex

Enums

Error

Functions

hex_check

Check if the input is valid hex bytes slice

hex_check_fallback

Check if the input is valid hex bytes slice

hex_check_sse⚠sse4.1

Safety

hex_check_sse_with_case⚠sse4.1

Safety

hex_check_with_case

Check if the input is valid hex bytes slice with case check

hex_decode

Hex decode src into dst. The length of src must be even and not zero. The length of dst must be at least src.len() / 2.

hex_decode_fallback

hex_decode_unchecked

hex_encode

Hex encode src into dst. The length of dst must be at least src.len() * 2.

hex_encode_fallback

hex_encode_upper

hex_encode_upper_fallback

hex_toDeprecated